move utilities and add test cases
