+++ art/poison_detection/provenance_defense.py
@@ -28,7 +28,7 @@ import numpy as np
 
 from art.poison_detection.ground_truth_evaluator import GroundTruthEvaluator
 from art.poison_detection.poison_filtering_defence import PoisonFilteringDefence
-from art.poison_detection.utils import segment_by_class, performance_diff
+from art.utils import segment_by_class, performance_diff
 
 from copy import deepcopy
 from sklearn.model_selection import train_test_split
